Fixes modifiers type signature in input_key callback (#276)
(tui) input_key event handler signature change (uint8_t to 16)

+++ src/shmif/arcan_tui.h
@@ -339,7 +339,7 @@ struct tui_cbcfg {
  *
  */
 	void (*input_key)(struct tui_context*, uint32_t symest,
-		uint8_t scancode, uint8_t mods, uint16_t subid, void* tag);
+		uint8_t scancode, uint16_t mods, uint16_t subid, void* tag);
 
 /*
  * other input- that wasn't handled in the other callbacks
